requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
ping3==4.0.4
requests==2.29.0
urllib3==1.26.15
Werkzeug==2.3.3
zipp==3.15.0

pika==1.3.2
python-dotenv==1.0.0
celery==5.2.7
gunicorn==20.1.0
flower==1.0.0
redis==4.1.4

Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.4